doctor_ai/requirement.txt
crewai
crewai-tools
browser_use
asyncio
python-dotenv
langchain_google_genai
Flask
gunicorn
